Sum of two numbers
First, a boolean variable validInput is declared to keep track of whether the user has entered valid input. Two do-while loops are used to repeatedly ask the user for input until valid input is received. Inside each loop, the user is prompted to enter a number. If the user enters a valid number, the validInput flag is set to true and the loop ends. If the user enters an invalid number, the program clears the input stream, ignores all remaining data until the next newline character and displays an error message, prompting the user to try again.

The numeric_limits<streamsize>::max() function is used to set the maximum number of characters that can be ignored by cin.ignore(), ensuring that any remaining input is properly cleared.

After the user has entered two correct numbers, they are added and the sum is displayed. This program is more complicated than the usual addition of two numbers because it implements error handling and input validation to make sure that the program works correctly, even if the user enters invalid data.